+++ libwfb/include/private/radiotap_iter.h
@@ -1,8 +1,8 @@
 #ifndef __RADIOTAP_ITER_H
 #define __RADIOTAP_ITER_H
 
+#include <private/radiotap_rc.h>
 #include <stdint.h>
-#include "radiotap_rc.h"
 
 /* Radiotap header iteration
  *   implemented in radiotap.c
@@ -10,11 +10,11 @@
 
 struct radiotap_override {
 	uint8_t field;
-	uint8_t align:4, size:4;
+	uint8_t align : 4, size : 4;
 };
 
 struct radiotap_align_size {
-	uint8_t align:4, size:4;
+	uint8_t align : 4, size : 4;
 };
 
 struct ieee80211_radiotap_namespace {
@@ -70,8 +70,8 @@ struct ieee80211_radiotap_iterator {
 	uint32_t *_next_bitmap;
 
 	unsigned char *this_arg;
-	const struct radiotap_override *overrides;	/* Only for RADIOTAP_SUPPORT_OVERRIDES */
-	int n_overrides;				/* Only for RADIOTAP_SUPPORT_OVERRIDES */
+	const struct radiotap_override *overrides; /* Only for RADIOTAP_SUPPORT_OVERRIDES */
+	int n_overrides;			   /* Only for RADIOTAP_SUPPORT_OVERRIDES */
 	int this_arg_index;
 	int this_arg_size;
 
@@ -83,12 +83,11 @@ struct ieee80211_radiotap_iterator {
 	int _reset_on_ext;
 };
 
-extern int ieee80211_radiotap_iterator_init(
-	struct ieee80211_radiotap_iterator *iterator,
-	struct ieee80211_radiotap_header *radiotap_header,
-	int max_length, const struct ieee80211_radiotap_vendor_namespaces *vns);
+extern int ieee80211_radiotap_iterator_init(struct ieee80211_radiotap_iterator *iterator,
+					    struct ieee80211_radiotap_header *radiotap_header,
+					    size_t max_length,
+					    const struct ieee80211_radiotap_vendor_namespaces *vns);
 
-extern int ieee80211_radiotap_iterator_next(
-	struct ieee80211_radiotap_iterator *iterator);
+extern int ieee80211_radiotap_iterator_next(struct ieee80211_radiotap_iterator *iterator);
 
 #endif /* __RADIOTAP_ITER_H */
